final(?) version
Working version that supports command line. You should type the .exe name in the cmd line and then type in three .txt file names which are the source of text and two output files. If any of the output files already exist there will be a dialog message shown which asks you if you want to rewrite files or not, for each output file.